+++ crates/mega-evm/tests/rex/main.rs
@@ -2,4 +2,3 @@
 
 mod intrinsic_gas;
 mod storage_gas;
-mod storage_gas_opcodes;
